Safari-611.1.10/LayoutTests/http/tests/security/contentSecurityPolicy/report-multiple-violations-01-expected.txt

CONSOLE MESSAGE: [Report Only] Refused to load the image 'http://127.0.0.1:8000/security/resources/abe.png' because it violates the following Content Security Policy directive: "img-src 'none'".

CONSOLE MESSAGE: [Report Only] Refused to load the image 'http://127.0.0.1:8000/security/resources/eba.png' because it violates the following Content Security Policy directive: "img-src 'none'".

PingLoader dispatched to 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/does-not-exist'.
This tests that multiple violations on a page trigger multiple reports. The test passes if two PingLoader callbacks are visible in the output.
